function_app/requirements.txt
azure-functions==1.18.0
azure-identity==1.17.1
azure-storage-blob==12.19.1
openai>=1.30.0
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1